+++ packages/yup/test/resolver.spec.ts
@@ -1,4 +1,12 @@
-import { collectNames, silk, weave } from "@gqloom/core"
+import {
+  collectNames,
+  field,
+  mutation,
+  query,
+  resolver,
+  silk,
+  weave,
+} from "@gqloom/core"
 import {
   GraphQLInt,
   GraphQLObjectType,
@@ -8,7 +16,7 @@ import {
 } from "graphql"
 import { assertType, describe, expect, expectTypeOf, it } from "vitest"
 import { type InferType, boolean, date, number, object, string } from "yup"
-import { field, mutation, query, resolver, union } from "../src"
+import { union, yupSilk } from "../src"
 
 describe("yup resolver", () => {
   const Giraffe = object({
@@ -33,8 +41,8 @@ describe("yup resolver", () => {
 
   collectNames({ Cat, Dog }, { Giraffe, GiraffeInput })
 
-  const createGiraffe = mutation(Giraffe, {
-    input: GiraffeInput,
+  const createGiraffe = mutation(yupSilk(Giraffe), {
+    input: yupSilk(GiraffeInput),
     resolve: (data) => ({
       name: data.name ?? "Giraffe",
       birthday: data.birthday ?? new Date(),
@@ -46,13 +54,13 @@ describe("yup resolver", () => {
     createGiraffe: createGiraffe,
   })
 
-  const giraffeResolver = resolver.of(Giraffe, {
-    age: field(number(), async (giraffe) => {
+  const giraffeResolver = resolver.of(yupSilk(Giraffe), {
+    age: field(yupSilk(number()), async (giraffe) => {
       return new Date().getFullYear() - giraffe.birthday.getFullYear()
     }),
 
-    giraffe: query(Giraffe, {
-      input: { name: string().required() },
+    giraffe: query(yupSilk(Giraffe), {
+      input: { name: yupSilk(string().required()) },
       resolve: ({ name }) => ({
         name,
         birthday: new Date(),
@@ -60,8 +68,8 @@ describe("yup resolver", () => {
       }),
     }),
 
-    greeting: field(string(), {
-      input: { myName: string() },
+    greeting: field(yupSilk(string()), {
+      input: { myName: yupSilk(string()) },
       resolve: (giraffe, { myName }) =>
         `Hello, ${myName ?? "my friend"}! My name is ${giraffe.name}.`,
     }),
@@ -132,13 +140,13 @@ describe("yup resolver", () => {
     collectNames({ Animal, Cat, Dog })
 
     const animalResolver = resolver({
-      cat: query(Animal, () => ({
+      cat: query(yupSilk(Animal), () => ({
         name: "Kitty",
         age: 1,
         loveFish: true,
       })),
 
-      dog: query(Animal, () => ({
+      dog: query(yupSilk(Animal), () => ({
         name: "Sadie",
         age: 2,
         loveBone: true,
@@ -223,12 +231,12 @@ describe("yup resolver", () => {
 
     const horseResolver = resolver.of(Horse, {
       createHorse,
-      hello: field(string(), (horse) => {
+      hello: field(yupSilk(string()), (horse) => {
         assertType<IHorse>(horse)
         return `Neh! Neh! --${horse.name}`
       }),
       horse: query(Horse, {
-        input: { name: string().required() },
+        input: { name: yupSilk(string().required()) },
         resolve: ({ name }) => ({
           name,
           age: 1,
@@ -290,7 +298,7 @@ describe("yup resolver", () => {
       })
 
       const animalResolver = resolver({
-        cat: query(Cat, () => ({
+        cat: query(yupSilk(Cat), () => ({
           name: "Kitty",
           age: 1,
         })),
@@ -315,12 +323,12 @@ describe("yup resolver", () => {
         age: number(),
       })
       const animalResolver = resolver({
-        cat: query(Cat, () => ({
+        cat: query(yupSilk(Cat), () => ({
           name: "Kitty",
           age: 1,
         })),
-        addCat: mutation(Cat, {
-          input: object({ data: Cat }),
+        addCat: mutation(yupSilk(Cat), {
+          input: yupSilk(object({ data: Cat })),
           resolve: ({ data }) => data,
         }),
       })